Extract shared viewer utilities to viewerUtils.ts
getLanguageFromPath (with its large languageMap) was duplicated in
MonacoViewer and MonacoDiffViewer. getMimeType and IMAGE_EXTENSIONS
were duplicated in ImageViewer and ImageDiffViewer. Centralise all
four into a single viewerUtils.ts file.
